PLANET-5870 Class per command
